fixup! Enh UUID validator
